fix: Fix off-by-one error in the computation of ScanP2Producer splits
The addition of `id` in the offsets vector in `compute_offsets` produces an `offsets` result
vector of size `data.len() + 1`. This wouldn't be an issue, as the final offset is ignored in `zip`,
if the length of that offset vector wasn't used to compute a midpoint split in `ScanP2Producer`.

This commit also introduces a failing test, but the length of the vector needed to exhibit the problem
may depend on the chunking logic of the local machine.